Only fold a symtable once
A statement can't tell if it owns a new symbol table or whether it's an
alias to an existing one, for example, an if-stmt is given a new
symtable when it has a symbol + initialiser. Thus fold may be called on
a symtable more than once, and we don't want to generate duplicate
symbols.