Attempt at fixing stack alignment in rare callee-save cases
Bug is present when callee save happens and a call takes place.
Callee-save alignment is done with regards to cc1_mstackalign, but not
to the alignment given in the impl_call() code. This means that the
impl_call() code thinks it's 16-byte align even though cc1_mstackalign
is 8, but then callee-save register saving comes along and realigns the
stack to 8 (cc1_mstackalign), messing this up.