Remove blocks - generate output exprs post-asm()
We need to generate the output expressions after the asm() inputs
and inlining. At this point we have the outputs in their temporaries,
e.g. "r" in a register, "m" in temporary memory, and we know what
registers have been allocated to meet this.

We can now generate output lvalues without fear of clobbering
anything as the backend takes care of reg spill normally, so we can
assign to the output lvalues easily.