Function inlining - convert rvalues to lvalues where necessary
This means if an inline function argument is addressed or written,
we create an alloca entry for it, otherwise we use its rvalue value
directly, which means functions with addressed/written arguments
can be inlined and functions with arguments used as pure rvalues
can take part in constant propagation without memory reads.

This could be seen as a mem2reg / mem2const optimisation that doesn't
use the backend directly.