Fix __builtin_va_start() va_list access expr
Now instead of having:
	va_list l;
	va_start(l); // l->x
we say
	va_list l;
	va_start(l); // (*l).x

which fits in with the more stringent array type system, since
(without decay), l is of type __builtin_va_list[1], which means to
assign we can't use ->, because without a correct type decay we try
to use -> on an array type, not pointer. Using dereference and '.'
is a simpler solution than decay the pointer to an array and keeping
lvalueness.