Ensure V_CONST_F out_vals are loaded in x86_fp_conv()
Inline functions may inline functions that do constant operations
on floats, e.g.

float add(float a){ return a + 1; }

so we want inline-float-calls to maintain the rvalueness of the float
argument when possible. But we also want the rvalue-float to convert
to a label + load (x86) in other cases, e.g.

void va_func(char *, ...);
float print(float f){ va_func("", f); }

Here we need to promote 'f' to double and this requires V_REG, not
V_CONST_F, so x86_fp_conv() now additionally tests for this.