lval2rval cast handles inline argument literal insertion
gen_expr_identifier() no longer does out_new_sym_val, as it's an
lvalue gen function, so it just emits an address for symbols.

This means out_new_sym_val() isn't called, so the symbol/immediate
value interposition optimisation for inline functions isn't done,
which meant the merge from master would've required all inline
functions to spill their arguments for a possible load.

expr_cast (specifically the lval2rval part) now catches this - if
it's about to lval2rval an identifier (or anything with a symbol),
and it can find a inline-immediate for said symbol, it will use
that instead.