Emit only reachable blocks - using two passes
The first pass figures out which blocks are reachable, the second
emits all reachable blocks in the right order.

This beats using one pass as we know if a block is reachable the
first time we emit it, ensuring the right block ordering, rather
than trying to emit it later in the wrong place when we figure out
that it is reachable.