Symtables may now be transparent - fixes C89 nested scope
Previously in C89 an if-statement shared scope with its surrounding
code block, meaning gen_block_decls() would be called twice for the
same scope in the same function (without inlining).

Now an if-statement in C89 has a separate scope for gen_block_decls(),
but it is "transparent", so used for looping over other constructs,
but decls (and typedefs, as required in C89) pass through to the
above scope.

This means the following still works in C89:

if((enum A { X, Y })0){ ... }
f(X); // lookup finds X in current scope