Fix memory access overflow from unused register in stret code
0 0x10b5b6a4a in v_unreserve_reg
1 0x10b5a7b1d in impl_overlay_mem_reg
2 0x10b5a7558 in impl_overlay_mem2regs
3 0x10b6068b9 in x86_func_ret_regs
4 0x10b6064d3 in impl_to_retreg
5 0x10b5b6e4f in out_ctrl_end_ret
6 0x10b5fb787 in gen_stmt_return
7 0x10b595c7a in gen_stmt
...

For small struct return, we would load the struct into a single
register, but still reserve and unreserve all registers that are
potentially used for stret. This meant that we would use a (potentially)
uninitialised value in regs[1].idx as an index into the reserved
registers array, e.g. octx->reserved_regs[regs[1].idx], which
wasn't causing problems as regs[...].idx is an unsigned short, so
unlikely to go over a page boundary, etc etc. ASAN caught this.