Fix extracting the numeric constant [lr]hs for const-folding ops
This fixes things like:

	5 == &a
	aka
	(T *)5 == &a

Now neither side is CONST_NUM, lhs is CONST_ADDR (numeric) and rhs
is CONST_ADDR (label). We were reading the offset of these and
thinking rhs' offset was zero, thinking it was a CONST_NUM of zero,
so allowing it as a compare-to-null constant expression.

test2/const_addr/op-global-bad.c caught this.