Spill handling - split apart stack/reg-mem values from spills
This splits apart the long-conflated confusion between how an lvalue and
spilt values are treated in the backend - lvalues are treated as
pointers to the location and we correctly dereference them from the
frontend (because of how lvalues work in the frontend with lvalue decay
"casts").

Spilt values were assumed to work like this, but we always want to
reload them on use, not just when they're explicitly dereferenced. This
is because the frontend (and other backend machinery) has no knowledge
of (and should have no knowledge of) a spilt value, which should just be
treated like a regular value, except that it's reloaded when used (i.e.
in v_to{,_reg}()).

This makes the separation clear and makes it harder to make mistakes
when dealing with potentially both types of value.