Fix argument addressing / addressing V_REGOFF in general
This fixes a bug exposed by 5e7ca35c01f9e72710f93391bee0a88d80768040,
with a test introduced (but disabled/ignored) in
d89573f1a17d1d58c716a5c817bdd096af19b63f.

A simpler repro to show that older commits also have the bug, is:

```
void g(int *);
void f(int a, float unused)
{
  g(&a);
}
```

The float forces the `mov` style argument saving code to be executed,
exposing the bug. The fix is to always treat `V_REGOFF` as an lvalue
and lea it, as we do with all other value types in the relevant switch
(except `V_SPILT`, which is already an indirect lvalue).

The test against `from->bits.regoff.offset` in the `V_REG` case was
almost always true (in the older versions of the code), which is why this
bug didn't show for `V_REG` values, and if it was false, we'd perform a
register copy, instead of running `impl_deref()`, which should only be
done for `V_SPILT`.