Fix sanitization of nonnull args
When the sanitizer code was committed on master, it was relying on
a code-gen bug that wouldn't load the address of arguments, hence
test/sanitize/nonnull-arg.c would pass.

This bug was fixed on the bootstrap branch in commit
ff3d809b759afb86d807ce44ee5b8b82fb90c5dd, and this commit fixes the
corresponding bug in the sanitization code.