Fix debug emission of unused locally-scoped global decls
The following tests now have locally-scoped decls, that don't exist in
the resultant binary because they're never reached:

  test/code_gen/dead_code_elim.c
  test/DCE/end.c

This means we attempted to emit debug information for them - normally we
only emit debug information for decls when generating the function (so
only if it's used) or not at all if it's implicit (as there's no
locally-scoped decl to cause the emission). This is a special case we
need to guard against debug emission to prevent link-time errors to
unemitted decls.

Note that we use decl_should_emit_code() to ensure we line up exactly
with what the code-gen considers as a function worth emitting, rather
than checking d->flags & DECL_FLAGS_USED/etc. This means that pure-inline
functions also won't be accidentally referenced by debug sections,
__attribute__((used)) will be respected on static functions, etc.