Fix handling of -mpreferred-stack-boundary=...
Previously, no -mpreferred-stack-boundary=... would mean we default to
log2() of the word size. So for 64-bit, this would mean 3. This value
would then be used in the code-base without being passed to pow(),
breaking the alignment algorithms, which assume a power-of-two argument.

When -mpreferred-stack-boundary was given, we'd correctly raise 2 to it,
meaning all the tests using the flag passed. The vla tests caught this
bug.