Emit constants requring relocation in a relro section
int *const p = &x;

The reference to `x` means `p` must be relocated at startup, which
means it can't live in a readonly section. Instead we place it in
a relro section, which is made readonly after relocations.

We could do this for all const decls, but instead we try to do it
just for those that require relocating.